+++ fairpy/course_allocation/algorithm3.py
@@ -121,4 +121,5 @@ def algorithm3(courses, students, students_matrix:list[list[bool]],csp_students:
 
 if __name__=="__main__":
     import pytest
-    pytest.main(args=["fairpy/course_allocation"])
+    #run algorithm and test of the algorithm
+    pytest.main(args=["fairpy/course_allocation/algorithm3.py", "fairpy/course_allocation/algorithm3_test.py"])
